coordinate vartype cleanup, i hope no new bugs :(

+++ pol-core/plib/filemapserver.cpp
@@ -26,14 +26,14 @@ FileMapServer::FileMapServer( const RealmDescriptor& descriptor ) :
     _cur_mapblock_index = 0;
 }
 
-MAPCELL FileMapServer::GetMapCell( unsigned x, unsigned y ) const
+MAPCELL FileMapServer::GetMapCell( unsigned short x, unsigned short y ) const
 {
     passert( x < _descriptor.width && y < _descriptor.height );
 
-    unsigned xblock = x >>  MAPBLOCK_SHIFT;
-    unsigned xcell  = x &   MAPBLOCK_CELLMASK;
-    unsigned yblock = y >>  MAPBLOCK_SHIFT;
-    unsigned ycell  = y &   MAPBLOCK_CELLMASK;
+    unsigned short xblock = x >>  MAPBLOCK_SHIFT;
+    unsigned short xcell  = x &   MAPBLOCK_CELLMASK;
+    unsigned short yblock = y >>  MAPBLOCK_SHIFT;
+    unsigned short ycell  = y &   MAPBLOCK_CELLMASK;
 
     long block_index = yblock * (_descriptor.width >> MAPBLOCK_SHIFT) + xblock;
     if (block_index != _cur_mapblock_index)
